chore(deps): upgrade Angular to 21.2.1 and fix compatibility issues
- Upgrade @angular/* packages from 17.3.12 to 21.2.1
- Update all Angular-related dependencies to compatible versions
- Fix component decorators and lifecycle hooks for Angular 21
- Update TypeScript configuration for Angular 21 compatibility
- Resolve breaking changes in Angular APIs
- All functionality tested and working normally

+++ tabby-terminal/src/features/zmodem.ts
@@ -143,7 +143,7 @@ class ZModemMiddleware extends SessionMiddleware {
                         if (canceled) {
                             return
                         }
-                        transfer.write(Buffer.from(chunk))
+                        transfer.write(Buffer.from(chunk) as Uint8Array)
                         this.showMessage(colors.bgYellow.black(' ' + Math.round(100 * transfer.getCompletedBytes() / details.size).toString().padStart(3, ' ') + '% ') + ' ' + details.name, true)
                     },
                 }),
